Add DeleteProfile component with confirmation popup and styles
